Support float coordinates in G1 (#95)

+++ main/motion.h
@@ -1,8 +1,8 @@
 #pragma once
 #include <Arduino.h>
 
-void moveAxis(int stepPin, int dirPin, long& pos, int target, int feedrate, char axis);
+void moveAxis(int stepPin, int dirPin, float& pos, float target, int feedrate, char axis);
 
 void homeAxis(int stepPin, int dirPin, int endstopPin, const char* label);
 
-void moveAxes(long targetX, long targetY, long targetZ, long targetE, int feedrate);
+void moveAxes(float targetX, float targetY, float targetZ, float targetE, int feedrate);
